Map generation and scene transition
Have made the map 0 (forest) and map 1 (mainroom of th house)
decided to use 2d vector for map.

map_generate_fail.cpp is the version where i use array to do, which failed lol

I also made the movement of map, but it is not a seperate file. Will do it later